+++ R/conjugacy.R
@@ -17,56 +17,62 @@
 #' @param beta the coefficients on \code{X}
 #' @param XtX,V0_inv,Xbeta pre-computed "shortcut" arguments for efficiency reasons
 #' @param use.chol Logical, determining whether to use \code{MASS::\link{mvrnorm}()} (\code{FALSE}, the default) or \code{\link{chol_mvrnorm}}.
+#' @param params.only Should just a list of the updated parameters be returned?
 #' @rdname conjugacy
 #' @export
-conj_norm_mu <- function(y, tau, mu0 = 0, tau0 = 0.001, mult = 1) {
+conj_norm_mu <- function(y, tau, mu0 = 0, tau0 = 0.001, mult = 1, params.only = FALSE) {
   n <- length(y)
   newtau <- mult*tau0 + n*tau
+  mean <- (tau0*mu0 + tau*sum(y)) / newtau
+  if(params.only) return(gu_params(mu = mean, tau = newtau, sd = 1 / sqrt(newtau)))
   stats::rnorm(
     1,
-    mean = (tau0*mu0 + tau*sum(y)) / newtau,
+    mean = mean,
     sd = 1 / sqrt(newtau)
   )
 }
 
 #' @rdname conjugacy
 #' @export
-conj_mvnorm_mu <- function(y, Q, mu0 = rep_len(0, p), Q0 = diag(0.001, p), mult = 1, use.chol = FALSE) {
+conj_mvnorm_mu <- function(y, Q, mu0 = rep_len(0, p), Q0 = diag(0.001, p), mult = 1, use.chol = FALSE, params.only = FALSE) {
   if(!is.matrix(y)) y <- matrix(y, nrow = 1)
   p <- ncol(y)
   n <- nrow(y)
   newQ <- mult*Q0 + n*Q
   newQ.inv <- chol_inv(newQ)
   mu <- newQ.inv %*% (Q0 %*% mu0 + Q %*% colSums(y))
+  if(params.only) return(gu_params(mu = mu, Q = newQ, Q.inv = newQ.inv))
   FUN <- if(use.chol) chol_mvrnorm else MASS::mvrnorm
   FUN(1, mu = mu, Sigma = newQ.inv)
 }
 
 #' @rdname conjugacy
 #' @export
-conj_matnorm_mu <- function(y, V, U = NULL, mu0, Q0, use.chol = FALSE) {
+conj_matnorm_mu <- function(y, V, U = NULL, mu0, Q0, use.chol = FALSE, params.only = FALSE) {
   if(!is.matrix(y)) stop("'y' must be a matrix")
   if(is.null(U)) U <- diag(nrow(Q0) / nrow(V))
   newQ <- V %x% U + Q0
   newQ.inv <- chol_inv(newQ)
   mu <- newQ.inv %*% (Q0 %*% mu0 + as.numeric(U %*% y %*% V))
+  if(params.only) return(gu_params(mu = mu, Q = newQ, Q.inv = newQ.inv))
   FUN <- if(use.chol) chol_mvrnorm else MASS::mvrnorm
   FUN(1, mu = mu, Sigma = newQ.inv)
 }
 
 #' @rdname conjugacy
 #' @export
-conj_lm_beta <- function(y, X, XtX = crossprod(X), tau, mu0, Q0, use.chol = FALSE) {
+conj_lm_beta <- function(y, X, XtX = crossprod(X), tau, mu0, Q0, use.chol = FALSE, params.only = FALSE) {
   newQ <- tau * XtX + Q0
   newQ.inv <- chol_inv(newQ)
   mu <- newQ.inv %*% (Q0 %*% mu0 + tau*t(X) %*% y)
+  if(params.only) return(gu_params(mu = mu, Q = newQ, Q.inv = newQ.inv))
   FUN <- if(use.chol) chol_mvrnorm else MASS::mvrnorm
   FUN(1, mu = mu, Sigma = newQ.inv)
 }
 
 #' @rdname conjugacy
 #' @export
-conj_matlm_beta <- function(y, X, V, U = NULL, mu0, Q0, use.chol = FALSE) {
+conj_matlm_beta <- function(y, X, V, U = NULL, mu0, Q0, use.chol = FALSE, params.only = FALSE) {
   if(!is.matrix(y)) stop("'y' must be a matrix")
   if(is.matrix(mu0)) mu0 <- as.numeric(mu0)
 
@@ -74,6 +80,7 @@ conj_matlm_beta <- function(y, X, V, U = NULL, mu0, Q0, use.chol = FALSE) {
   newQ <- V %x% (XtU %*% X) + Q0
   newQ.inv <- chol_inv(newQ)
   mu <- newQ.inv %*% (Q0 %*% mu0 + as.numeric(XtU %*% y %*% V))
+  if(params.only) return(gu_params(mu = mu, Q = newQ, Q.inv = newQ.inv))
   FUN <- if(use.chol) chol_mvrnorm else MASS::mvrnorm
   FUN(1, mu = mu, Sigma = newQ.inv)
 }
@@ -92,17 +99,16 @@ conj_matlm_beta <- function(y, X, V, U = NULL, mu0, Q0, use.chol = FALSE) {
 
 #' @rdname conjugacy
 #' @export
-conj_norm_tau <- function(y, mu, a0 = 0.001, b0 = 0.001) {
-  stats::rgamma(
-    1,
-    a0 + 0.5*length(y),
-    b0 + 0.5*sum((y - mu)^2)
-  )
+conj_norm_tau <- function(y, mu, a0 = 0.001, b0 = 0.001, params.only = FALSE) {
+  a <- a0 + 0.5*length(y)
+  b <- b0 + 0.5*sum((y - mu)^2)
+  if(params.only) return(gu_params(a = a, b = b))
+  stats::rgamma(1, a, b)
 }
 
 #' @rdname conjugacy
 #' @export
-conj_mvnorm_Q <- function(y, mu, V0, v0, V0_inv = chol_inv(V0)) {
+conj_mvnorm_Q <- function(y, mu, V0, v0, V0_inv = chol_inv(V0), params.only = FALSE) {
   if(!is.matrix(y)) y <- matrix(y, nrow = 1)
   p <- ncol(y)
   n <- nrow(y)
@@ -116,6 +122,7 @@ conj_mvnorm_Q <- function(y, mu, V0, v0, V0_inv = chol_inv(V0)) {
     t(txmu) %*% txmu
   }))
   V2 <- chol_inv(V0_inv + tmp)
+  if(params.only) return(gu_params(V = V2, v = n + v0))
   stats::rWishart(
     1,
     Sigma = V2,
@@ -125,7 +132,7 @@ conj_mvnorm_Q <- function(y, mu, V0, v0, V0_inv = chol_inv(V0)) {
 
 #' @rdname conjugacy
 #' @export
-conj_matnorm_V <- function(y, mu, U = NULL, V0, v0, V0_inv = chol_inv(V0)) {
+conj_matnorm_V <- function(y, mu, U = NULL, V0, v0, V0_inv = chol_inv(V0), params.only = FALSE) {
   if(!is.matrix(y) || !is.matrix(mu)) stop("'y' and 'mu' must be matrices")
   n <- nrow(mu)
   stopifnot(identical(dim(y), dim(mu)))
@@ -133,6 +140,7 @@ conj_matnorm_V <- function(y, mu, U = NULL, V0, v0, V0_inv = chol_inv(V0)) {
   ymu2 <- if(is.null(U)) t(ymu) else crossprod(ymu, U)
 
   V2 <- chol_inv(V0_inv + ymu2 %*% ymu)
+  if(params.only) return(gu_params(V = V2, v = n + v0))
   stats::rWishart(
     1,
     Sigma = V2,
@@ -142,14 +150,14 @@ conj_matnorm_V <- function(y, mu, U = NULL, V0, v0, V0_inv = chol_inv(V0)) {
 
 #' @rdname conjugacy
 #' @export
-conj_lm_tau <- function(y, X, beta, Xbeta = X %*% beta, a0 = 0.001, b0 = 0.001) {
-  conj_norm_tau(y = y, mu = Xbeta, a0 = a0, b0 = b0)
+conj_lm_tau <- function(y, X, beta, Xbeta = X %*% beta, a0 = 0.001, b0 = 0.001, params.only = FALSE) {
+  conj_norm_tau(y = y, mu = Xbeta, a0 = a0, b0 = b0, params.only = params.only)
 }
 
 #' @rdname conjugacy
 #' @export
-conj_matlm_sigma <- function(y, X, beta, Xbeta = X %*% beta, U = NULL, V0, v0, V0_inv = chol_inv(V0)) {
-  conj_matnorm_V(y = y, mu = Xbeta, U = U, v0 = v0, V0_inv = V0_inv)
+conj_matlm_sigma <- function(y, X, beta, Xbeta = X %*% beta, U = NULL, V0, v0, V0_inv = chol_inv(V0), params.only = FALSE) {
+  conj_matnorm_V(y = y, mu = Xbeta, U = U, v0 = v0, V0_inv = V0_inv, params.only = params.only)
 }